Security Fixes (All 9 fixed): - SEC-01: Encrypted seal storage (AES-256-GCM + Argon2id) wired into production - SEC-02: Nostr P2P transport with encrypted DM mode (NIP-04/44) - SEC-03: Nostr key persistence to ~/.csv/nostr_secret_key.hex - SEC-04: Chain configs no longer contain hardcoded API keys - SEC-05: Encrypted seal storage on wasm32 (IndexedDB) - SEC-06: Per-chain adaptive polling (±20% jitter) - SEC-07: ML-DSA-65 (Dilithium3) post-quantum signatures implemented

+++ csv-bitcoin/src/seal.rs
@@ -4,6 +4,7 @@ use crate::error::{BitcoinError, BitcoinResult};
 use crate::types::BitcoinSealPoint;
 use crate::wallet::Bip86Path;
 use csv_core::hardening::{BoundedQueue, MAX_SEAL_NULLIFIER_SIZE};
+use csv_core::store::SealStore;
 
 #[cfg(feature = "rpc")]
 use csv_store::SqliteSealStore;
@@ -86,7 +87,7 @@ impl SealRegistry {
         seal: &BitcoinSealPoint,
         height: u64,
     ) -> BitcoinResult<()> {
-        use csv_store::SealStore;
+        use csv_core::store::SealStore;
         use csv_core::{SealRecord, Hash};
         
         // First mark in memory
@@ -102,7 +103,7 @@ impl SealRegistry {
                 recorded_at: std::time::SystemTime::now()
                     .duration_since(std::time::UNIX_EPOCH)
                     .unwrap_or_default()
-                    .as_secs() as i64,
+                    .as_secs(),
             };
             // Use interior mutability via the trait method
             // Since we can't mutably borrow from &self, we need a different approach
